feat: rename project to "Lard Lad Exchange"
Swaps the generic "Hello Flink" title for a name that fits the subject:
DONU/DonutCoin transactions flowing through the pipeline, matched to Lard
Lad Donuts, Springfield's giant-donut landmark from The Simpsons.

Renames the br.avcaliani.hello_flink package to
br.avcaliani.lard_lad_exchange, the Flink job names, Kafka consumer
group id and transactional id prefix, and the run.sh banner, so the new
name shows up in the Flink dashboard and Kafka, not just the README.
Adds settings.gradle.kts to make rootProject.name explicit (previously
unset, silently defaulting to the directory name) so the built uber jar
is actually named lard-lad-exchange-*-uber.jar, matching run.sh's
JAR_FILE.

Also adds a .github/PULL_REQUEST_TEMPLATE.md matching the one in
deep-dive-pyspark.
