+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,0 +1,11 @@
+### 🔍 What?
+<!-- One or two sentences describing what this PR does. -->
+
+### 💡 Why?
+<!-- The motivation — what are you learning or solving? -->
+
+### 🔀 What's changing?
+
+| File(s) | Change |
+|---|---|
+| `` |  |
