+++ src/main/java/br/avcaliani/lard_lad_exchange/infra/serializers/KafkaSerializer.java
@@ -1,4 +1,4 @@
-package br.avcaliani.hello_flink.infra.serializers;
+package br.avcaliani.lard_lad_exchange.infra.serializers;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
